reorg folder and file convention, and add more details of the tests
